+++ app/Http/Livewire/Managment/DistributerLicencePregled.php
@@ -11,6 +11,20 @@ class DistributerLicencePregled extends Component
     public $distId;
     public $dist_name;
 
+    public $dataClicked;
+    public $wkey = 1;
+
+    protected $listeners = ['chartClicked' => 'getClickedData'];
+
+    public function getClickedData($data)
+    {
+        $this->wkey ++;
+        $this->dataClicked = $data;
+        $this->render();
+        //dd($data);
+        //$this->dispatchBrowserEvent('chartClicked', ['data' => $data]);
+    }
+
     /**
      * mount
      *
@@ -21,10 +35,13 @@ class DistributerLicencePregled extends Component
         $this->distId = request()->query('id') ?? 0;
 
         $this->dist_name = LicencaDistributerTip::where('id', '=', $this->distId)->first()->distributer_naziv ?? 'Sve licence';
+        $this->dataClicked = '';
     }
 
     public function render()
     {
-        return view('livewire.managment.distributer-licence-pregled');
+        return view('livewire.managment.distributer-licence-pregled', [
+            'dataDisp' => $this->dataClicked,
+        ]);
     }
 }