+++ resources/views/livewire/managment/distributer-licence-pregled.blade.php
@@ -5,8 +5,10 @@
         <div class="text-red-600 float-left">{{ $dist_name }}</div>
         </div>
     </div>
-
-    @livewire('managment.pocetak-licence-grafik', ['distId' => $distId, 'grafikNaslov' => 'Zeta naplata'])
+    
+    <livewire:managment.pocetak-licence-grafik :distId="$distId" />    
+    
+    <livewire:managment.prikaz-izabranih-licenci :dataClicked="$dataDisp" :distId="$distId" :wire:key="$wkey" />
     
 
 </div>